+++ src/main/java/com/ironSync/Main.java
@@ -1,15 +1,13 @@
 package com.ironSync;
 
 import com.ironSync.controller.UserController;
-import com.ironSync.model.WorkoutEntry;
-import com.ironSync.storage.FileManager;
+import com.ironSync.model.Workout;
 
 public class Main {
-    private static final FileManager fileHandler = new FileManager();
     public static void main(String[] args) {
         UserController userController = new UserController();
-        //System.out.println(userController.createNewWorkoutEntry());
-        WorkoutEntry createdWorkoutEntry = userController.createNewWorkoutEntry();
-        fileHandler.registerData(createdWorkoutEntry);
+        Workout createdWorkoutEntry = userController.createNewWorkout();
+
+        System.out.println(createdWorkoutEntry);
     }
 }