+++ src/main/java/com/ironSync/controller/UserController.java
@@ -7,12 +7,14 @@ import com.ironSync.util.UserInputs;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 /**
  * Controller class responsible for handling user input related to workouts and workout entries.
  * It interacts with utility classes to gather data and creates domain objects accordingly.
  */
 public class UserController {
+    private final Scanner scanner = new Scanner(System.in);
     private final UserInputs userInputs = new UserInputs();
     private final ParseDate dateParser = new ParseDate();
 
@@ -26,7 +28,7 @@ public class UserController {
         String exerciseDone = userInputs.promptText("Introduce the name of the exercise done:");
         List<Integer> numberOfReps = userInputs.repetitionsStringToIntegerList(
                 userInputs.promptText("Introduce the number of repetitions done:"));
-        int numberOfSets = userInputs.promptInt("Introduce the number of sets done:");
+        int numberOfSets = numberOfReps.size();
 
         return new WorkoutEntryController().workoutEntryBuilder(exerciseDone, numberOfReps, numberOfSets);
     }
@@ -42,10 +44,11 @@ public class UserController {
         while (true) {
             workoutEntries.add(createNewWorkoutEntry());
 
-            System.out.println("Do you want to add another workout entry? ");
-            String userDecision = userInputs.promptTextWithOptions("Do you want to add another workout entry? (Y/n)", "y,n");
+            userInputs.promptTextWithOptions("Do you want to add another workout entry? (Y/n)", "y,n");
 
-            if (userInputs.processUserInputFromTwoOptions("y,n", userDecision)) {
+            String userDecision = scanner.nextLine();
+
+            if (!userInputs.processUserInputFromTwoOptions("y,n", userDecision)) {
                 return workoutEntries;
             }
         }
@@ -56,12 +59,13 @@ public class UserController {
      * a workout title, and workout notes as entered by the user.
      * The current date and time is automatically assigned to the workout.
      *
-     * @param workoutEntries list of WorkoutEntry objects to include in the workout
      * @return a new Workout object with the specified data
      */
-    public Workout createNewWorkout(List<WorkoutEntry> workoutEntries) {
+    public Workout createNewWorkout() {
+
         String workoutTitle = userInputs.promptText("Introduce the name of the workout: ");
         String workoutNotes = userInputs.promptText("Introduce the notes for the workout: ");
+        List<WorkoutEntry> workoutEntries = this.buildWorkoutEntryList();
 
         return new WorkoutController().workoutBuilder(workoutEntries, workoutTitle, workoutNotes, dateParser.getCurrentDateTimeFormatted(1));
     }